+++ src/main/java/com/sparta/petnexus/post/postComment/dto/PostCommentRequestsDto.java
@@ -1,7 +1,7 @@
-package com.sparta.petnexus.postComment.dto;
+package com.sparta.petnexus.post.postComment.dto;
 
 import com.sparta.petnexus.post.entity.Post;
-import com.sparta.petnexus.postComment.entity.PostComment;
+import com.sparta.petnexus.post.postComment.entity.PostComment;
 import com.sparta.petnexus.user.entity.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
