+++ src/main/java/com/sparta/petnexus/post/postComment/repository/PostCommentRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.petnexus.postComment.repository;
+package com.sparta.petnexus.post.postComment.repository;
 
-import com.sparta.petnexus.postComment.entity.PostComment;
+import com.sparta.petnexus.post.postComment.entity.PostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostCommentRepository extends JpaRepository<PostComment, Long> {
