+++ src/main/java/com/sparta/petnexus/post/postLike/repository/PostLikeRepository.java
@@ -1,7 +1,7 @@
-package com.sparta.petnexus.postLike.repository;
+package com.sparta.petnexus.post.postLike.repository;
 
 import com.sparta.petnexus.post.entity.Post;
-import com.sparta.petnexus.postLike.entity.PostLike;
+import com.sparta.petnexus.post.postLike.entity.PostLike;
 import com.sparta.petnexus.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
